+++ designer/DesignerApp.ts
@@ -39,12 +39,12 @@ class DesignerCommander extends AbstractCommander {
     this._app._onGrammarKeyup()
   }
 
-  simulateDataCommand() {
+  synthesizeProgramCommand() {
     const grammarProgram = new jtree.GrammarProgram(this._app.getGrammarCode())
     this._app.setCodeCode(
       grammarProgram
         ._getRootNodeTypeDefinitionNode()
-        .generateSimulatedData()
+        .synthesizeNode()
         .join("\n")
     )
     this._app._onCodeKeyUp()
@@ -538,8 +538,8 @@ class tableComponent extends AbstractTreeComponent {
    a Download Bundle
     ${WillowConstants.DataShadowEvents.onClickCommand} downloadBundleCommand
    span  |
-   a Generate Random Program
-    ${WillowConstants.DataShadowEvents.onClickCommand} simulateDataCommand
+   a Synthesize Program
+    ${WillowConstants.DataShadowEvents.onClickCommand} synthesizeProgramCommand
    textarea
     id grammarConsole
   td